Add proper Y offset detection to DetectSwipe
Pick p0 and pN from the list of points, p0 being the start point
with left most x value and pN being the end point with right most
x.

Build a function (calcF) and calculate for each point between
p0 and pN the Y value. If the actual Y value of the point differs
too much from the calculated one, it's not a swipe.